LAst commit 2

+++ Data/Models/DataPlan.cs
@@ -5,8 +5,8 @@ namespace Data.Models
     public class DataPlan
     {
         public int DataPlanId { get; set; }
-        public int NationalData { get; set; }
-        public int InternationalData { get; set; }
+        public float NationalData { get; set; }
+        public float InternationalData { get; set; }
         public List<MobilePhoneDataPlanAssignment> MobilePhoneDataPlanAssignments { get; set; }
     }
 }
