+++ Data/Models/MobilePhoneCall.cs
@@ -7,7 +7,7 @@ namespace Data.Models
         public int PhoneNumber { get; set; }
         public PhoneLine PhoneLine { get; set; }
         
-        public int IMEI { get; set; }
+        public long IMEI { get; set; }
         public MobilePhone MobilePhone { get; set; }
         
         public DateTime DateTime { get; set; }
