+++ Data/Models/MobilePhoneEmployee.cs
@@ -2,7 +2,7 @@ namespace Data.Models
 {
     public class MobilePhoneEmployee
     {
-        public int IMEI { get; set; }
+        public long IMEI { get; set; }
         public  MobilePhone MobilePhone { get; set; }
         
         public  int EmployeeId { get; set; }
